pyproject.toml
[project]
name = "balance-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "numpy",
    "gspread",
    "oauth2client",
    "ipykernel",
    "python-dotenv",
    "reflex",
    "streamlit",
    "ace_tools"
]
